In this topic, we've defined the class ComplexNumber and several magic methods for this class. However, it makes sense to define a couple of other methods so that we can use this class to its full potential.
We need methods for subtraction and division.

Subtraction

If we have 2 complex numbers x = a + bix=a+bi and y = c + diy=c+di , then x - yx−y will equal (a - c) + (b - d)i(a−c)+(b−d)i.

Division

For division, we need to calculate the reciprocal of a complex number. If x = a + bix=a+bi, then {1 \over x} = {a \over a^2 + b^2} - {b \over a^2 + b^2}i
x
1
​
 =
a
2
 +b
2

a
​
 −
a
2
 +b
2

b
​
 i.

{x \over y}
y
x
​
 can then be represented as x * {1 \over y}x∗
y
1
​
 .